Parse track tokens greedily
The previous code stored the state of the parsing in a regex string.
However, that allowed for situations where a pattern that made sense
would get more tokens appended to it, because the regular expression
itself didn't keep the information which characters were already covered
by the previously appended regular expression fragments.
The new system keeps track of the remainders of each row and only
checks the new tokens on the remainders, thus providing a greedy way of
covering the track list by a regular expression.